Change form field to Decimal

+++ forms.py
@@ -6,4 +6,4 @@ from .models import AccountUser
 class TransferForm(forms.Form):
     user_from = forms.ModelChoiceField(queryset=AccountUser.objects.all(), empty_label='От кого')
     inn_to = forms.IntegerField(label='Кому')
-    amount = forms.FloatField()
+    amount = forms.DecimalField()
